Ok big update, so here goes some explanation.
1) I removed the khasbot package from starcraft.khasbot and moved it to starcraft.proxybot.khasbot Hopefully i did it
right in the repository and the remove will delete out the files for you guys, if not then i'll double check this on
my windows machine after this commit.

2) ProxyBot is no longer an agent. We had it all wrong. I forget how many hours i spent trying to figure out why the
behaviours i was adding never executed, hence the birth of ProxyBotAgent. ProxyBotAgent is the go between for the regular
java application (ProxyBot) and the JADE platform. All communication will go through ProxyBotAgent and passed to the
commander. This will break how it runs in Eclipse, but not really sure. Like I said tired and will test later ( i know
bad commit, but i want to make sure the code is there and it not get lost)

3) I have some very basic communication going from the ProxyBot.java application to ProxyBotAgent.java. I need to finish
the other side of it passing information from JADE platform to Java application.

4) ProxyBot.java now creates the ProxyBotAgent on JADE

5) ProxyBotAgent now creates the CommanderAgent on JADE

6) CommanderAgent now creates the remaining agents on JADE ( currently commented out )

7) .......

8) Profit